Backend/requirements.txt
keras
flask
flask_cors
pillow
numpy
base64